feat: local project support for numbers projects
This commit represents a re-write of the support for numbers
projects. It has two objectives:
- enable storing training data in indexeddb storage
  (This is optional, and storing training data in cloud PgSQL
   remains an option)
- move running the model to the browser
  (This is required - when this goes live, models will no longer
   be hosted in the cloud.)

Models will still be created in the cloud, but they will only be
stored for a very short period. Browsers will need to download
and store the model locally.

This will represent a major reduction in cloud workload, as
all model inferencing for non-text projects will now be moved off
of the site servers.

+++ mlforkids-api/public/components/projects/projects.service.js
@@ -40,11 +40,16 @@
             }
         }
 
-        function getFields(projectid, userid, tenant) {
-            return $http.get('/api/classes/' + tenant + '/students/' + userid + '/projects/' + projectid + '/fields')
-                .then(function (resp) {
-                    return resp.data;
-                });
+        function getFields(project, userid, tenant) {
+            if (project.storage === 'local') {
+                return Promise.resolve(project.fields);
+            }
+            else {
+                return $http.get('/api/classes/' + tenant + '/students/' + userid + '/projects/' + project.id + '/fields')
+                    .then(function (resp) {
+                        return resp.data;
+                    });
+            }
         }
 
         function getLabels(project, userid, tenant) {
@@ -168,7 +173,7 @@
 
         function deleteProject(project, userid, tenant) {
             if (project.storage === 'local') {
-                if (project.cloudid) {
+                if (project.type === 'text' && project.cloudid) {
                     $http.delete('/api/classes/' + tenant + '/students/' + userid + '/localprojects/' + project.cloudid);
                 }
 
@@ -236,13 +241,14 @@
                 return {
                     scratch : true,
                     edublocks : true,
-                    python : true
+                    python : true,
+                    appinventor : true
                 };
             }
             else if (project.type === 'numbers') {
                 return {
                     scratch : true,
-                    edublocks : true,
+                    edublocks : false,
                     python : true
                 };
             }
@@ -265,7 +271,8 @@
                     return {
                         scratch : true,
                         edublocks : false,
-                        python : true
+                        python : true,
+                        appinventor : true
                     };
                 }
             }
